+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 ### Features
 
 * bem.scoped
+
+## v2.0.0
+
+### Features
+
+* The package is now an ESM module.
+* `bem.unscoped` and `bem.single` was deprecated.
+* `bem.scoped` is now just `bem` - the default export.
+* Added the `bemStyles` function to make BEM classnames from CSS module styles.